+++ src/main/java/nu/superserver/timereport/db/Month.java
@@ -1,10 +1,10 @@
 package nu.superserver.timereport.db;
 
 public class Month {
-    private final long year;
+    private final int year;
     private final java.time.Month month;
 
-    public Month(long year, java.time.Month month) {
+    public Month(int year, java.time.Month month) {
         this.year = year;
         this.month = month;
     }
@@ -13,7 +13,7 @@ public class Month {
         return month;
     }
 
-    public long getYear() {
+    public int getYear() {
         return year;
     }
 }
